ParallelFor for Reduction (#1658)
Add capability for `ParallelFor` to safely do reduction using `deviceReduceSum`,
`Min`, etc.  The user passes `Gpu::KernelInfo{}.setReduction(true)` to notify
`ParallelFor` that this is a parallel reduction, and gives `ParallelFor` a callable
that takes `Gpu::Handler`.   A `Gpu::Handler` is needed to call
 `deviceReduceSum`.

Also add `Gpu::Buffer` class, whose data pointer can be used as a device
destination for `deviceReduceSum`.  It also has a `copyToHost` method to
copy the device result back to the host. 

See `Tutorials/GPU/ParallelReduce` for examples of how to use `ParallelFor`
for reduction.

Also note that the reduction function is OpenMP CPU threads safe.  Thus the
same code can run on with OpenMP when it is not built for GPU.

Co-authored-by: Andrew Myers <atmyers2@gmail.com>