Always have Abort_device etc. defined for GPU build
It used to be these functions are defined only if NDEBUG is defined.
However, the cmake build system does not store `NDEBUG` in `AMReX_Config.H.
It's possible that AMReX is built with `NDEBUG` and `Abort_device` is not
defined, but the user's code may be built without `NDEBUG`.  Then
`amrex::Abort` defined in `AMReX.H` would fail to compile.  Always having
these device functions defined can solve the issue.