RoundRobin: option to not sort processes
Add an optional parameter `sort` to RoundRobin constructors.  This makes it
consistent with other DistributionMapping strategies.  And this fixes an
incorrect behavior when RoundRobin is used by KnapSack that has been given
`sort=false`.  This should be able to fix the hanging issue observed in
WarpX simulations with load balancing enabled and more MPI processes than
the number of boxes.