Refinement Ratio of 4 Support in Nodal Solver (#1774)
* Refinement Ratio of 4 Support in Nodal Solver

This adds refinement ratio of 4 support in non-EB nodal linear solver.
Roundoff errors are expected for multi-level solves with a refinement ratio
of 2.

* MacProjector refinement ratio of 4 bug

Move setLevelBC to project and call it only if it has not been called.  This
fixes a refinement ratio of 4 bug, because setLevelBC should not be called
before coarse/fine ratio is set.