The_Async_Arena and Elixir::append (#1804)
Add The_Async_Arena that uses CUDA steam-ordered memory allocator, if available.  If the
feature is not available, use The_Arena to allocate and free memory and the free is done
with a host callback function.  The purpose of this Arena is to be used by a temporary
FArrayBox inside an MFIter loop.

Add Elixir::append that appends an rvalue Elixir to an existing one to reduce the number of
host callback calls.