AMREX_NO_PROBINIT
For Amr/AmrLevel based codes, there is now a compile time option
AMREX_NO_PROBINIT=[TRUE|FALSE] for the support of the probin file and
amrex_probinit function.  The default behavior of requiring a user defined
amrex_probinit function unless amr.probin_file is empty does not change
unless amrex is built into a library.  The reason for the inconsistency is
there is no portable way of supporting a weak symbol (.e.g, undefined
amrex_probinit in amrex itself) in a shared library.

This will partially solve the shared library build issue on Mac and
Windows.