Arena Release Threshold
Release memory from Arena back to the system if it's above a threshold.  For
the pinned arena, the release threshold is set to the size of the global
memory.  For others, it's set to a huge number.  These can be controlled
with ParmParse parameters, `amrex.the_arena_release_threshold` and
`the_[device|pinned|managed|async]_release_threshold`.  The reason for
setting the pinned arena's release threshold to the size of the global
memory is that we received a report in the past that the pinned arena used
by the communication functions wasted a huge amount of memory in the case
that more and more grids were added over the run time.