LevelBld (#1852)
In Amr class, we need to get LevelBld* and use it to build AmrLevels.
Previously, Amr called a function `LevelBld* getLevelBld()` to get that.
The problem was that `getLevelBld` was not defined by amrex but it was being
called by amrex.  This caused issues in building amrex as a shared library
on Mac and Windows, because they do not support weak symbols.

In this PR, the constructors of Amr now take `LevelBld*` as an argument.
This is unfortunately a breaking change.  However, it's a compile time
failure and it's easy to fix.  See the changes in this commit for examples.