Add FabArray::release and Make BaseFab virtual
Add FabArray::release functions that release the ownership of the FAB at
the given index and return a raw pointer to the FAB.

Make BaseFab virtual so that it's safe to use the new release functions.

Also declare a number of functions `[[nodiscard]]`.

Fix some warnings in EB/CNS example.