Fix a bug in BaseFab::resize.
Previously if BaseFab::resize's optional `Arena*` argument is nullptr (which
is the default) while BaseFab's Arena* member is a different Arena, the
nullptr will be interpreted as the default Arena, `The_Arena`, ignoring the
BaseFab's Arena has already been set.  This causes the following code to
allocate data using The_Arena, not The_Cpu_Arena.

    BaseFab fab(The_Cpu_Arena());
    fab.resize(box);

This PR fixes the bug.