CPU version of TagBox::buffer
Add a CPU version of TagBox::buffer.  @MSABuschman reported in #1951 that
TagBox::buffer has been very slow since commit #1258 if the error buffer
size is large.  The function was rewritten in #1258 to do the work on GPU.
In this PR, the old version is reintroduced for CPU.

Note that the current implementation is expected to have poor performance on
GPU if it has a very large error buffer.  It's still not clear how we should
implement this function for GPU if a large error buffer is used.