Undeprecate ReduceData::value()
ReduceData::value() with no argument was recently marked as a depracated
function because it involves calling a std::function.  However, the call is
never made inside a loop like MFIter.  So there is no practical difference
in performance between ReduceData::vaule and ReduceData::value(ReduceOps&).
It is hereby undeprecated to avoid warning messages for a lot of
applications.