Fix proper nesting issue for large error buffer (#1987)
Reimplement the code that ensures proper nesting by projecting the fine
grids onto the coarse.  The previous implementation has a bug that causes it
to fail when the error buffer is very large.

In the very beginning of BoxLib, the enforcement of proper nesting was
actually done in a similar way as the approach in this commit (i.e., calling
TagBoxArray::setVal(BoxArray,TagBox::CLEAR) with projected down BoxArray).
However, it had the issue of buffering the buffer cells, creating
unnecessarily big refinement patches.  That issue was fixed in a block of
very complicated code that was hard to reason, also in the very beginning of
BoxLib days.  Since it worked, the code had not been touched at all.
However, it was reported recently it failed when a very large error buffet
was used.  In this commit, we switched back to the original setVal approach
and it is performed after the tagged cells are buffered already, thus
avoiding the issue of buffering the buffer cells.  Furthermore, it is
performed after the tags are coarsened, thus performing less work.

This also adds a check to make sure that blocking factors do not vary too
much between levels.  Otherwise regrid will break.