Add new parameter to BoxArray::contains(BoxArray)
In regrid, we need to figure out how many cells we need to grow the current
coarse level grids in order for the fine grids to be properly nested.  We
used BoxArray::contains(BoxArray) in a loop and grew the coarse BoxArray in
the loop.  In this commit, we add a new parameter to
BoxArray::contains(BoxArray) to specify the number of grow cells without
having to actually grow the BoxArray.  This is better than the previous
version because growing BoxArray would invalidate the hash data in BoxArray.
We also modify the implementation of BoxArray::contains(Box) in this commit.