Cray complier type detection
There are two types of Cray compilers, the classical and clang based.  The version number
is not a reliable way to determine the type.  For example, cce/9 on Cori has both
versions.  Cray compilers for aarch64 on Ookami are classical.  Thus we switch to grep
"LLVM" in `CC --version`.  This works on both Cori and Ookami.