ParallelFor for MultiFab/FabArray (#2073)
Add an experimental feature that can be used to launch a single kernel to
work on MultiFabs or FabArrays.  An example is shown below, where `mf` is a
MultiFab.
```
        auto ma = mf.arrays();
        amrex::ParallelFor(mf,
                           [=] AMREX_GPU_DEVICE (int box_no, int i, int j, int k)
        {
            ma[box_no](i,j,k) = 1.0;
        });
        Gpu::synchronize();
```

It's still unclear how we handle tiling and OpenMP and how it performs on CPU.  However, there is strong evidence that this improves performance for GPU runs when there are small boxes.  The purpose of this PR is to get the feature into the AMReX repo first and do further tuning later.  Since the new function is  not used at all, it should not affect any codes.