CMake 3.20+ CUDA: Shell-Escape --Werror
When adding a list of flags to CMake targets, duplicated flags get
by default deduplicated. This caused with our new CMake 3.20+ CUDA
logic that this gets passed to the linker:
`--Werror ext-lambda-captures-this cross-execution-space-call`.

The CMake `SHELL:` prefix is exactly made for this instance, keeping
flags that belong together also over transformations like this
together.

Since we injected directly into the global `CMAKE_CUDA_FLAGS` flags
for earlier CMake versions, this deduplication logic does not
apply and thus we have not experienced it earlier.