Enhance DiffSameGrid2.cpp (#2114)
This utility, when using L0 norm, now outputs an additional column with the absolute error divided by the maximum value on the coarse grid.  This is useful for comparing/verifying data with local regions of very small (close to zero in a machine precision sense) values that may fluctuate based on roundoff errors.