IParser: Integer Parser
Add a parser for integers.  Because there is rounding in integer division,
it's not safe to use the real number parser for integer.

The two parsers have a lot of similarity, but floating point number specific
functions (e.g., `sqrt`, `sin`, etc.) are not supported in `IParser`.  In
addition to `/` whose result truncates towards zero, the integer parser also
supports `//` whose result truncates towards negative infinity.