Fix lincc_interp
This fixes a bug in lincc_interp that was introduced more than 2 years ago
when the Fortran code was converted to C++.  The slope was mistakenly
limited twice.  For each component, we compute a limiting factor and we use
the minimum of all components as the final limiting factor.  That final
limiting factor should be applied to the unlimited central difference slope,
not the monotonized central slope.