EB: Fix small cells and multicuts (#2225)
The previous approach of fix small cells and multicuts is not consistent.
If a small cell is turned into a covered cell, we need to fix its neighbors
too.  In this new approach, we build the EB data iteratively.  This commit
also introduces a new ParmParse parameter, eb2.maxiter, that controls the
maximal number of iterations.