fix bug in AmrLevel exclusion logic (#2283)
## Summary
Fixes a bug in the build logic regarding optional inclusion of the AMRLEVEL components, and return a required  dependency to AMReX_AmrParticleDataAdaptor.H

## Additional background
This bug was introduced in https://github.com/AMReX-Codes/amrex/pull/2258 where the SENSEI adaptors and bridges that processed data from the Amr class. A key dependency was removed, in the refactor. This Pr adds that dependency back in, and moves `AMReX_AmrParticleInSituBridge.H` behind the `AMReX_AMRLEVEL` compile option where it belongs.