Remove some versions of ParallelFor(MF) (#2326)
Remove ParallelFor(MF) functions that only take `int ncomp`, but not `IntVect nghost` to
avoid potential bugs. The versions we have kept are,

 * ParallelFor(MF const&, ...); // valid region only
 * ParallelFor(MF const&, IntVect const& nghost, ...); // valid+nghost
 * ParallelFor(MF const&, IntVect const& nghost, int ncomp, ...); // valid+nghost and loop over component

The version that is removed is

  * ParallelFor(MF const&, int ncomp, ...);

This avoids bugs like below

    ParallelFor(mf, mf.nGrow(), ...); // should use mf.nGrowVect()

Here, `mf.nGrow()` returns `int` and therefore would be treated as the number of
components in the removed version.