CUDA: -G by Default
In the past, this often did not compile at all, was very sensitive to
further set options, or compiled super slowly;
in some cases, such as recursive function usage, apps need to increase
`cudaLimitStackSize` in order to not stack overflow with device debug
symbols (this costs some extra DRAM).

Nonetheless, for CUDA approx. 11.0+, we see the opposite: we have
very slow Debug builds with CUDA if we do not activate -G for some
`LinearSolvers/MLMG` objects. Thus, we default-on now to -G in Debug
builds.