CMake CUDA Debug: -G by Default (#2355)
## Summary

In the past, this often did not compile at all, was very sensitive to further set options, or compiled super slowly; in some cases, such as recursive function usage, apps need to increase `cudaLimitStackSize` in order to not stack overflow with device debug symbols (this costs some extra DRAM).

Nonetheless, for CUDA approx. 11.0+, we see the opposite: we have very slow Debug builds with CUDA if we do not activate -G for some `LinearSolvers/MLMG` objects: `AMReX_MLNodeLaplacian.cpp` & `AMReX_MLNodeLaplacian_sten.cpp`. Thus, we default-on now to -G in Debug builds.

- [x] testing: now `LinearSolvers/MLMG/AMReX_MLTensorOp.cpp` compiles a bit longer (1-2min), but that's better than the other two TUs above without `-G` in Debug

## Additional background

Thanks to @WeiqunZhang and @maxpkatz to looking into this with me.