Fix read unitialized memory issue in MLMG solver (#2373)
## Summary
Running valgrind on one of our failing circleCi tests revealed that the MLMG solver read allocated but uninitialized memory. Setting `m_phi_eb` to 0 after it is allocated fixes this issue. Attached is the valgrind logs, Line 17275 shows where valgrind detected reading of uninitialized memory.

[vgout.txt](https://github.com/AMReX-Codes/amrex/files/7269235/vgout.txt)

## Additional background
This was found due to one of our warpx circleCI test failing after a fetch and merge with development. This fix did not solve that issue though.