Avoid free after amrex::Finalize
Previously calling Arena::free after amrex::Finalize resulted in segfault
because the Arena objects have been deleted.  It will cause the following
code to crash in the end.

    int main (int argc, char* argv[])
    {
        amrex::Initialize(argc, argv);
        amrex::MultiFab(...);
        amrex::Finalize();
    }

One way to fix it is to add a scope to the function body between Initialize
and Finalize so that the object inside will go out of scope before
amrex::Finalize is called.  In this commit, we implement a way of avoiding
this kind of errors even if one forgets to free everything allocated by
Arenas before Finalize.