Default of the Arena's Release Threshold
Change the default of the Arena's release threshold from max (i.e., no
release) to the maximum of the initial allocation size and 3/4 of the GPU
memory.  This helps the memory oversubscription issue observed in some WarpX
runs.  Once we have to allocate additional device memory because the initial
allocation, which is 3/4 of the total global memory, has been used up, the
new additions to the Arena are not big enough for efficient reuse, resulting
in fragmentation in the Arena.  Because at that point we are very close to
use up all device memory, it's much more important to make sure we do not
oversubscribe due to memory fragmentation by releasing memory back to the
system than the performance gain we might have in memory allocation beyond
the initialize size.