Add Filesystem::DirectorySeparator
Add a public function to that returns the platform specific directory
separator. We currently need this in WarpX and defining this in a
central place makes sense for re-use.

It looks like plotfile writing works on Windows as of now through
std filesystem operations, so no further changes are needed at this
point unless we observe otherwise down the road.