Port CellQuadratic to C++ with GPU support (#2369)
## Summary
The CellQuadratic interpolater has been ported from Fortran to C++ with GPU support.
The errors for the CPU-only version on local are in the ~1e-15 range for AmrLevel SingleVortex 2D, ~1e-15 for IAMR 2D bubble Cartesian, and ~1e-12 for IAMR 2D bubble RZ.
The errors for the GPU-enabled version on Gigan are in the ~1e-14 range for AmrLevel SingleVortex2D, ~1e-10 for IAMR 2D bubble Cartesian, and ~1e-9 for IAMR 2D bubble RZ.

## Additional background
This port is heavily influenced by the CellConservativeLinear interpolater. It reuses two of the same slope kernels.
With this PR, _all_ of the interpolaters now work on the GPU!
Also, thanks to @cgilet for helping with debugging.