Remove DArena and EArena (#2426)
DArena that uses the buddy memory allocation strategy and EArena that uses
the best fit strategy are experimental features we explored in the past.
But they have never been used because they do not work as well as the first
bit based CArena.  There is no reason to keep them anymore.