Fix use-after-free (#2443)
The parallel copy meta-data cache is store in std::multimap<BDKey,CPC*>.  A
CPC* might be stored multiple times in the multimap, and we need to delete
the pointers before clearing the multimap.  The previous way of avoiding
double free was incorrect, because it had a use-after-free.  In this commit,
we fix it by pushing a unique copy of the pointers to a vector first while
iterating the multimap.

This bug only manifests if a user's code has a memory leak of
MultiFab/FabArray.