AMREX_RELEASE_NUMBER
Add a new macro `AMREX_RELEASE_NUMBER`.  It is a six digits number.  The
first two are for the year, and the next two are for the month.  The last
two are usually 00, unless a patch other than the monthly release is
released.  If the AMReX source is a git version that's not the release
version, the last release number will be used.

For git version obtained with git --describe, we define something like

    #define AMREX_GIT_VERSION "21.10-65-gfbb2edc95551-dirty"
    #define AMREX_RELEASE_NUMBER 211000

For git version obtained from the CHANGES file, we define something like

    #define AMREX_GIT_VERSION "21.10"
    #define AMREX_RELEASE_NUMBER 211000

or

    #define AMREX_GIT_VERSION "21.10.1"
    #define AMREX_RELEASE_NUMBER 2110001

or

    #define AMREX_GIT_VERSION "21.10.11"
    #define AMREX_RELEASE_NUMBER 211011