amrex_particles_to_vtp.py samplefiles fixed
amrex_particles_to_vtp.py skips the first four lines after the first
line (containing the number of particles) of the input file. Then, the
number of lines that are read is equal to exactly the number of
particles specified on the first line.

The sample input files under the samplefiles/ directory do not contain
these four "skipped" lines, and as a result, not only does
amrex_particles_to_vtp.py skip four lines of valid particle data, the
script crashes because it tries to read past the end of the file. (In
other words, the script assumes that the file is four lines longer than
it really is.)

As a workaround, I've added four "dummy" lines to each file. This is NOT
an ideal solution because real particle output would not contain my four
"dummy" lines. However, my workaround is good enough for the purpose of
demonstrating correct usage of amrex_particles_to_vtp.py. I have
confirmed that the resulting .vtp files produce sensible-looking results
in ParaView.