Use specific targets for make uninstall (#2471)
This updates the `make uninstall` rule which is used with configure to more specifically remove targets which were installed. It also adds a dependency file removal for the `make cleanconfig` rule. This removes the assumption that the directory set from `./configure --prefix=/my/example/directory` is empty, and therefore all files should be removed with `make uninstall`

Co-authored-by: Weiqun Zhang <WeiqunZhang@lbl.gov>