Mapped solver support (#2088)
This change to the AMReX_MLNodeLaplacian* files allows us to pass in a multi-component sigma to the nodal projector and this will be used as {sigmax,sigmay,sigmaz} internally.  This allows us to have different weightings for the derivatives in the different coordinate directions which is useful for solves on nonuniform rectilinear meshes.