ParmPaser::queryAdd
Add a new member function, queryAdd, to ParmParse.  If the name is found,
the value in the ParmParse database will be stored in the reference
argument.  If not, the value in the argument will be added to the ParmParse
database.  The return value indicates if it existed previously.

Many codes use ParmParse::dumpTable to store the ParmParse data in
plot/checkpoint files.  But for parameters that not present in inputs files
or command line arguments, they would not appear in the dumped tables.  The
new queryAdd function can be used to address this issue.

In amrex/Src, the calls to `query` are replaced by `queryAdd` so that it's
easy to find their values.