TableData: A Multi-Dimensional Array Class
This class is somewhat similar to FArrayBox/BaseFab. The main difference is
the dimension of the array in this class can be 1, 2, 3, or 4, whereas the
dimension of FArrayBox/BaseFab is the spatial dimension (AMREX_SPACEDIM)
plus a component dimension.  This is a feature request by Maui, which will
use it to store EOS data.