SUNDIALS MRI Interface for AMReX TimeIntegrator class (#2600)
This PR adds an interface to the SUNDIALS ARKODE integrator for the AMReX TimeIntegrator class to use, identical to what we have implemented for the ERF code. This will let us move that implementation into the more abstract AMReX time integration API to allow others to take advantage of it.

The new capabilities are described in the Docs page for Time Integration.

This capability is a work in progress and we will add features as we need them.