Allow a descriptor component's BndryFuncFabDefault to be stateful. (#2612)
## Summary

This allows a boundary condition function to depend on the descriptor index or other information only available when BndryFunc::setComponent() is called.

## Additional background

Previously, the BndryFuncFabDefault function passed to BndryFunc::setComponent() pointed to a function that did not take descriptor index or other information. In some cases (e.g. multi-material simulations where each material is in its own Descriptor), extra information is needed to map the boundary-function call to the correct material's boundary condition.
Making BndryFuncFabDefault a std::function instead of a plain function pointer makes this possible.