Introduce `AMReX_Version.H`
The defined `AMREX_GIT_VERSION` and `AMREX_RELEASE_NUMBER` change in
every commit. This modifies the `AMReX_Config.H` file as well for
every commit, which is included in every header file.

Using always the latest commit counteracts `ccache`
since the signature of the compilation call changes every commit.

Move this define and `AMREX_RELEASE_NUMBER` to a separate version
file, which is included only where needed (currently in a single
`.cpp` file).