Semicoarsening in WarpX Linear Solvers
Add support for semicoarsening in two linear solvers used by WarpX,
MLNodeTensorLap and EBNodeFDLaplacian.  The semicoarsening approach here is
different from the existing approach for the nodal projection solver.  There
the grids are coarsened regularly as much as possible and then we switch to
coarsening in directions that are still coarsenable, whereas here we do
semicoarsening first in specified direction and then switch to regular
coarsening if the specified maximum semicoarsening level has not been
reached.