FillBoundaryAndSync (#2683)
Add a new function to FabArray that fills ghost cells and synchronize nodal
data. Ghost regions are filled with data from the intersecting valid
regions. The synchronization will override valid regions by the intersecting
valid regions with a higher precedence.  The smaller the global box index
is, the higher precedence the box has.  With periodic boundaries, for cells
in the same box, those near the lower corner have higher precedence than
those near the upper corner.  This function produces the same result as
OverrideSync followed by FillBoundary in a single pass.

Also modify the implementation of OverrideSync to use this new function,
which is more efficient.

The nodal linear solvers use OverrideSync to sync up the nodal data. They
have been updated to use the new OverrideSync function that does not use
an owner mask. So we no longer need to build owner masks except for the
top and bottom multigrid levels that need to use the mask in the computation
of norms to avoid overcounting.