FillPatchUtil edge interpolater fix (#2701)
This PR addresses a compilation error that occurs when using templated FabArrays that are not MultiFab. (The static_cast of the abstract mapper Interpolater prevented compilation with derived and templated Interpolater objects.) To fix this, a templated InterpFace function was introduced that can be used by derived Interpolater, but that provides default behavior with casting for regular Interpolaters. (Note: this is similar to what is done with FillPatchInterp).