Fix FluxRegister::SumReg for GPU builds (#2718)
It's incorrect to assume that the fused version of reduce is only used on
FabArrays with a simple BoxArray.  FluxRegister::SumReg uses the fused
version of reduce, and its BoxArray has a special BATransformer.  A
FluxRegister only has cells on the faces of regular Boxes, but the meta-data
provided to the GPU reduce function is for regular Boxes.  This results in
out-of-bound errors.  This commit fixes the issue by extending the meta-data
caching mechanism to include special BoxArrays used by FluxRegisters.

We have also refactored Batransformer's opeartor==.