Reimplement amrex::min and max to Work around an nvcc bug
This is a workaround for an nvcc bug when compiling ERF in debug mode.

Note that because std::min and max have been constexpr functions since C++14
and AMReX uses C++14 now, there is no need to use amrex::min and max
anymore.  But for backward compatibility, we will still keep them.