Fix a bug in multigrid grids
This fixes a bug introduced in #2690. We have to make sure every box in the
fine BoxArray is coarsenable, otherwise average down functions will fail.
For example, for a 10x10 cells domain broken into four 5x5 cells boxes, we
should not add a coarse level because 5x5 is not coarsenable, even though
the 10x10 domain is coarsenable. In principle, we could fix this in the
average down functions instead. However, these tests with odd numbers of
cells are corner cases.  So we are fixing it in a relatively easy approach.