Add: `MultiFab::sum_unique`
This provides a new method to sum values in a `MultiFab`.
For non-cell-centered data, `MultiFab::sum` double counts box
boundary values that are owned by multiple boxes. This provides
a function that does not double count these and provides a
quick way to get only the sum of physically unique values.