Fix Tensor Solver BC
This fixes some bugs in the physical domain BC of tensor linear solver.

At the corner of two no-slip walls (e.g., (0,0)), we have u(-1,0) = -u(0,0)
and u(0,-1) = -u(0,0).  It's incorrect to fill the corner ghost cell with
u(-1,-1) = u(-1,0) + u(0,-1) - u(0,0), because it will result in u(-1,-1) =
-3 * u(0,0).

In the old approach, to avoid branches in computing transverse derivatives
on cell faces, we fill the ghost cells first.  For example, to compute du/dy
at the lo-x boundary, we use the data in i = -1 and 0, just like we compute
du/dy(i) using u(i-1) and u(i) for interior faces.  The problem is the
normal velocity in the ghost cells outside a wall is filled with
extrapolation of the Dirichlet value (which is zero) and more than 1
interior cells. Because of the high-order extrapolation, u(-1) != -u(0).
This is the desired approach for computing du/dx on the wall.  However, this
produces incorrect results in dudy.

In the new approach, we explicitly handle the boundaries in the derivative
stencil.  For example, to compute transverse derivatives on an inflow face,
we use the boundary values directly.