Fix EB data inconsistency when fixing small cells and multiple cuts
For consistency, we need to call the function that zeros out the level set
even if that box does not have any small cells or multiple cuts.  This is
because a node could exist in multiple boxes.  Furthermore, a covered cell
or covered face may have a node with a level set < 0.