ParallelFor with compile time optimization of kernels with run time parameters
Branches inside ParallelFor can be very expensive.  If a branch uses a lot
of resources (e.g., registers), it can significantly affect the performance
even if at run time the branch is never executed because it affects the GPU
occupancy.  For CPUs, it can affect vectorization of the kernel.

The new ParallelFor functions use C++17 fold expression to generate kernel
launches for all run time variants.  The kernel function can use constexpr
if to discard unused code blocks for better run time performance.  Here are
two examples of how to use them.

    int runtime_option = ...;
    enum All_options : int { A0, A1, A2, A3};
    // Four ParallelFors will be generated.
    ParallelFor(TypeList<CompileTimeOptions<A0,A1,A2,A3>>{},
                {runtime_option},
                box, [=] AMREX_GPU_DEVICE (int i, int j, int k, auto control)
    {
        ...
        if constexpr (control.value == A0) {
            ...
        } else if constexpr (control.value == A1) {
            ...
        } else if constexpr (control.value == A2) {
            ...
        else {
            ...
        }
        ...
    });

and

    int A_runtime_option = ...;
    int B_runtime_option = ...;
    enum A_options : int { A0, A1, A2, A3};
    enum B_options : int { B0, B1 };
    // 4*2=8 ParallelFors will be generated.
    ParallelFor(TypeList<CompileTimeOptions<A0,A1,A2,A3>,
                         CompileTimeOptions<B0,B1> > {},
                {A_runtime_option, B_runtime_option},
                N, [=] AMREX_GPU_DEVICE (int i, auto A_control, auto B_control)
    {
        ...
        if constexpr (A_control.value == A0) {
            ...
        } else if constexpr (A_control.value == A1) {
            ...
        } else if constexpr (A_control.value == A2) {
            ...
        else {
            ...
        }
        if constexpr (A_control.value != A3 && B_control.value == B1) {
            ...
        }
        ...
    });

Note that that due to a limitation of CUDA's extended device lambda, the
constexpr if block cannot be the one that captures a variable first.  If
nvcc complains about it, you will have to manually capture it outside
constexpr if.  The data type for the parameters is int.

Thank Maikel Nadolski and Alex Sinn for showing us the meta-programming
techniques used here.